backend/llm_service/pyproject.toml
[tool.poetry]
name = "copilot-llm-service"
version = "0.1.0"
description = "LLM микросервис для обработки запросов с поддержкой файлов"
authors = ["Аслан Алибеков <aslan.alibekov.2017@mail.ru>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.ruff]
target-version = "py313"
line-length = 120
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    "alembic/versions",
]

[tool.ruff.lint]
preview = true
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "S",
    "BLE",
    "T20",
    "PT",
    "RUF",
    "ERA",
    "PL",
    "PERF",
]
fixable = ["ALL"]
unfixable = []
ignore = [
    "ERA001",
    "RUF001",
    "B008",
    "F821",
]

[tool.ruff.lint.pylint]
max-args = 8
max-positional-args = 8

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = false
lines-after-imports = 2
known-first-party = ["app"]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]

[tool.mypy]
python_version = "3.13"
files = ["app"]
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
show_error_codes = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.121.1"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = "^2.0.0"
pydantic-settings = "^2.12.0"
httpx = "^0.28.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
asyncpg = "^0.30.0"
alembic = "^1.17.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
pymupdf = "^1.25.5"
python-docx = "^1.1.2"
chardet = "^5.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
